requirements.txt
requests==2.28.2
black==23.1.0
python-decouple==3.7